pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "^1.26.1"
pandas = "^2.1.2"
python-dotenv = "^1.0.0"
feedparser = "^6.0.10"
pysqlite3 = "^0.5.2"
sqlalchemy = "^2.0.22"
google-api-python-client = "^2.105.0"
croniter = "^2.0.1"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.26.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# logging setup
[logging]
version = 1
disable_existing_loggers = false

[logging.formatters.simple]
format = "%(asctime)s %(name)s:%(lineno)s %(funcName)s [%(levelname)s]: %(message)s"

[logging.handlers.consoleHandler]
class = "logging.StreamHandler"
level = "INFO"
formatter = "simple"
stream = "ext://sys.stdout"

[logging.handlers.fileHandler]
class = "logging.FileHandler"
level = "INFO"
formatter = "simple"

[logging.loggers.__main__]
level = "DEBUG"
handlers = ["consoleHandler", "fileHandler"]
propagate = false

[logging.loggers.same_hierarchy]
level = "DEBUG"
handlers = ["consoleHandler", "fileHandler"]
propagate = false

[logging.loggers.lower.sub]
level = "DEBUG"
handlers = ["consoleHandler", "fileHandler"]
propagate = false

[logging.root]
level = "INFO"